feat(ai): tiered AI chat infrastructure (Plan A) (#85)
* refactor(ui): rename typo_graphy.dart to app_typography.dart

* refactor(ui): move AppColor into lib/core/theme/

* refactor(ui): move ThemeProvider into lib/core/theme/

* fix(billing): guard billingService behind Platform.isAndroid to prevent iOS crash

* fix(ui): remove ServiceLocator from _UnlockProBanner — route through MainViewModel


* fix(pro): replace addPostFrameCallback in build + typed error keys in PaywallViewModel

Convert _PaywallContent to StatefulWidget so the isPurchased listener fires context.pop() exactly once via initState, eliminating spurious pops on subsequent rebuilds. Replace the fragile errorMessage.startsWith('pro.') heuristic with typed errorKey/errorArgs fields so BillingError platform messages never trigger bogus translation lookups.


* chore(i18n): remove unused billing keys error_network and restored

* test(billing): migrate billing service test to patrolTest wrapper

* docs(roadmap): add unified enhancement roadmap spec


* refactor(search): split 875-LOC search_content into widget part files

Extracts search bar, sort, filter panel, filter chips, result card, and
empty/header state widgets into lib/views/search/widgets/ part files per
File_Organization_Rules R1. Each file now under the 300 LOC hard limit.


* refactor(ai): split 681-LOC rag_service into retriever/attribution/support

Recipe R2: extract RagRetriever (rewrite→expand→embed→search→RRF→resolve),
RagAttribution (source titles/debug notes), RagSupport (messages, error
mapping, chat-memory prep, debug formatting), and RagResult into their own
files. RagService stays the orchestrator at 294 LOC. Public API unchanged —
RagService.rrfFuse kept as a static delegate; constructor + ServiceLocator
getter signatures preserved. All RAG/chat/quiz tests pass.


* refactor(ai): make LlmClient a facade over LlmChatBackend providers

Finishes the in-progress provider migration. LlmClient (667->191 LOC) now
selects and delegates to GeminiApiLlmProvider / OpenAiCompatibleLlmProvider
in initialize(), converting LlmChatMessage->ChatTurn. Removes the duplicate
inline LlmApiException in favour of the canonical llm_api_exception.dart
(re-exported for callers). Ports the Gemini model-resolution + streaming
fallback into GeminiApiLlmProvider via a new GeminiModelResolver helper.

ServiceLocator wiring and public API unchanged; defaultGeminiModel regression
lock preserved. All llm/rag/chat/quiz tests pass; analyze clean.


* refactor(notes): slim NoteViewModel 534->248 LOC

Removes dead toolbar code (toggleBold/isBold/_updateToolbarState/
clearAllFormatting etc.) — the editor uses flutter_quill's QuillSimpleToolbar,
which drives formatting through the controller, so these had zero callers.
Extracts the active 'Enter inherits inline formatting' engine into
NoteEditorKeyHandler and content parsing into NoteDocumentCodec (Recipe R3),
and DRYs the four identical tag-setters. Public VM API unchanged; analyze clean.


* refactor(tags): split 580-LOC UnifiedTagsIconButton into part files

Extracts the dialog (_UnifiedTagsDialog), category sidebar (_CategorySidebar),
generic chip (_TagChip), and per-category selection pane (_TagTabBody) into
widgets/ part files. DRYs the four identical selection handlers into one
_toggle helper. Public UnifiedTagsIconButton API unchanged; each file now under
the 300 LOC limit; analyze clean.


* refactor(tags): split 513-LOC custom_tags_widget into separate files

- Move CompactCustomTagsWidget to compact_custom_tags_widget.dart
- Extract chip widgets (_ExistingTagChip, _StringTagChip, _CustomTagChip)
  into part file widgets/custom_tag_chips.dart
- custom_tags_widget.dart re-exports CompactCustomTagsWidget for back-compat
- custom_tags_widget.dart reduced from 513 to ~230 LOC


* feat(billing): add IosStorekitBillingService behind IBillingService

Wire trovara/billing MethodChannel for iOS (StoreKit).
Platform.isIOS branch selects the new impl; non-Android/iOS
falls back to StubBillingService. Native Swift handler is a
follow-up PR.


* docs(ai): spec tiered AI chat (free on-device + Pro premium cloud)

Expands roadmap Tier 4.1. Free tier = on-device LLM (zero marginal
cost, unlimited, offline) with optional BYOK; Pro = premium cloud +
enhanced retrieval. No RAG rewrite — tiers the generation backend and
retrieval depth behind the existing LlmChatBackend seam.


* docs(ai): plan A tiered AI chat infrastructure (pure Dart)

7 TDD tasks: ChatTierResolver, RetrievalDepth presets threaded into
RagService, OnDeviceLlmProvider stub, ByokKeyStore, ServiceLocator
tier wiring, chat tier badge, depth-driven chat send. On-device
runtime + model packaging deferred to Plan B.


* feat(pro): Tier 1.2/1.3 billing interface, Android service, paywall route

Stage the reviewed Tier 1 deliverables: IBillingService + Android Play
billing service, PaywallView + /pro/paywall route, phase2 plans and Pro
specs, and the objectbox note repo cleanup.


* docs(ai): fix plan A test_support import paths + callback param


* feat(ai): add ChatTier + ChatTierResolver for tiered chat


* feat(ai): tier-aware retrieval depth in RagService/RagRetriever


* feat(ai): add OnDeviceLlmProvider stub + LlmProvider.onDevice


* feat(ai): add ByokKeyStore for free-tier bring-your-own-key


* test(ai): assert free tier skips multi-query expansion

Resolves Task 2 review Minor[T2a]: lock the RetrievalDepth.free
expansion-skip behavior (rewrite runs, expand does not).


* feat(di): wire tier-aware chat backend + retrieval depth


* feat(chat): tier badge + free-tier upgrade nudge


* feat(chat): drive RAG retrieval depth from active tier


---------

+++ lib/core/repository/implementations/objectbox_note_repository.dart
@@ -1,12 +1,11 @@
 import 'package:logger/logger.dart';
+// Import ServiceLocator for deferred graph analysis
+import 'package:trovara/core/di/service_locator.dart';
 import 'package:trovara/core/repository/base/base_repository.dart';
 import 'package:trovara/core/repository/base/objectbox_store_manager.dart';
 import 'package:trovara/core/repository/interfaces/note_repository.dart';
 import 'package:trovara/models/note.dart';
 import 'package:trovara/objectbox.g.dart';
-
-// Import ServiceLocator for deferred graph analysis
-import 'package:trovara/core/di/service_locator.dart';
 
 /// ObjectBox implementation of the note repository.
 ///
@@ -241,7 +240,7 @@ class ObjectBoxNoteRepository extends BaseRepository implements INoteRepository 
     // Fire and forget
     try {
       final knowledgeGraphService = ServiceLocator().knowledgeGraphService;
-      knowledgeGraphService.analyzeNote(note.id as int, note.contentJson);
+      knowledgeGraphService.analyzeNote(note.id, note.contentJson);
     } catch (e) {
       _logger.w('Failed to queue graph analysis for note ${note.id}', error: e);
     }
